// src/app/pages/appointment-management/history-medical-appointments/models/history-medical-appoiment.ts
export class HistoryMedicalAppoimentFilter {
    idPersonaDTO: string;
    estadoCitaDTO: string;
    especialidadDTO: string;
    instalacionPrestadorDTO: string;
    fechaInicialDTO: Date;
    fechaFinalDTO: Date;

}

export class HistoryMedicalAppoimentDTO {
    estadoDTO: string;
    fechaCitaDTO: string;
    horaCitaDTO: string;
    instalacionPrestadorDTO: string;
    tipoCitaDTO: string;
    especialidadMedicaDTO: string;
    nombreEspecialitaDTO: string;
    apellidoEspecialitaDTO: string;
    consultorioDTO: string;
    procedimientoMedicoDTO: string;
}